requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
pydantic
python-dotenv
matplotlib
scipy
